@@ -14,8 +14,8 @@ publish_to: none
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
-  flutter: ">=3.16.0"
+  sdk: '>=3.4.0 <4.0.0'
+  flutter: ">=3.22.0"
 dependencies:
 
   # The following adds the Cupertino Icons font to your application.
@@ -25,16 +25,18 @@ dependencies:
   cupertino_icons: ^1.0.4
 
   ff_annotation_route_library: ^3.0.0
-  extended_text: ^12.0.0
+  extended_text: ^14.0.1
   # extended_text:  
   #   version: ^11.0.0-dev.1
   #   hosted: "https://pub.dev"  
   flutter:
     sdk: flutter
   html: ^0.15.0
-  loading_more_list: ^4.1.0
+  loading_more_list: ^7.1.0
   oktoast: ^3.1.5
   url_launcher: ^6.0.17
+  extended_keyboard: ^0.0.3
+  extended_image: any
 
 dev_dependencies:
   flutter_test:
@@ -42,6 +44,7 @@ dev_dependencies:
 dependency_overrides:
   extended_text_field:
     path: ../  
+ 
   #extended_text:
     # git: 
     #   url: https://github.com/fluttercandies/extended_text.git
