@@ -1,17 +1,18 @@
 import 'package:flutter/material.dart';
 
 class ToggleButton extends StatefulWidget {
-  const ToggleButton(
-      {this.activeWidget,
-      this.unActiveWidget,
-      this.activeChanged,
-      this.active = false});
-  final Widget? activeWidget;
-  final Widget? unActiveWidget;
+  const ToggleButton({
+    this.activeChanged,
+    this.active = false,
+    super.key,
+    required this.builder,
+  });
+
   final bool active;
   final ValueChanged<bool>? activeChanged;
+  final Widget Function(bool active) builder;
   @override
-  _ToggleButtonState createState() => _ToggleButtonState();
+  State<ToggleButton> createState() => _ToggleButtonState();
 }
 
 class _ToggleButtonState extends State<ToggleButton> {
@@ -41,7 +42,7 @@ class _ToggleButtonState extends State<ToggleButton> {
       },
       child: Padding(
         padding: const EdgeInsets.only(left: 20.0, top: 10.0, bottom: 10.0),
-        child: widget.active ? widget.activeWidget : widget.unActiveWidget,
+        child: widget.builder(_active),
       ),
     );
   }
