@@ -140,7 +140,7 @@ class _ExtendedSelectableTextState extends _SelectableTextState {
     assert(debugCheckHasDirectionality(context));
     assert(
       !(widget.style != null &&
-          widget.style!.inherit == false &&
+          !widget.style!.inherit &&
           (widget.style!.fontSize == null ||
               widget.style!.textBaseline == null)),
       'inherit false style must supply fontSize and textBaseline',
