@@ -29,6 +29,7 @@ part 'package:extended_text_field/src/official/material/selectable_text.dart';
 class ExtendedTextField extends _TextField {
   const ExtendedTextField({
     super.key,
+    super.groupId = EditableText,
     super.controller,
     super.focusNode,
     super.undoController,
@@ -49,6 +50,7 @@ class ExtendedTextField extends _TextField {
     super.toolbarOptions,
     super.showCursor,
     super.autofocus = false,
+    super.statesController,
     super.obscuringCharacter = '•',
     super.obscureText = false,
     super.autocorrect = true,
@@ -71,6 +73,7 @@ class ExtendedTextField extends _TextField {
     super.cursorRadius,
     super.cursorOpacityAnimates,
     super.cursorColor,
+    super.cursorErrorColor,
     super.selectionHeightStyle = ui.BoxHeightStyle.tight,
     super.selectionWidthStyle = ui.BoxWidthStyle.tight,
     super.keyboardAppearance,
@@ -79,6 +82,7 @@ class ExtendedTextField extends _TextField {
     super.enableInteractiveSelection,
     super.selectionControls,
     super.onTap,
+    super.onTapAlwaysCalled = false,
     super.onTapOutside,
     super.mouseCursor,
     super.buildCounter,
@@ -99,11 +103,7 @@ class ExtendedTextField extends _TextField {
     this.extendedSpellCheckConfiguration,
     this.specialTextSpanBuilder,
     super.magnifierConfiguration,
-    this.onPasteFunction,
-
   });
-
-  final Function()? onPasteFunction;
 
   /// build your ccustom text span
   final SpecialTextSpanBuilder? specialTextSpanBuilder;
@@ -290,7 +290,7 @@ class ExtendedTextFieldState extends _TextFieldState {
     assert(debugCheckHasDirectionality(context));
     assert(
       !(widget.style != null &&
-          widget.style!.inherit == false &&
+          !widget.style!.inherit &&
           (widget.style!.fontSize == null ||
               widget.style!.textBaseline == null)),
       'inherit false style must supply fontSize and textBaseline',
@@ -299,10 +299,12 @@ class ExtendedTextFieldState extends _TextFieldState {
     final ThemeData theme = Theme.of(context);
     final DefaultSelectionStyle selectionStyle =
         DefaultSelectionStyle.of(context);
+    final TextStyle? providedStyle =
+        MaterialStateProperty.resolveAs(widget.style, _statesController.value);
     final TextStyle style = _getInputStyleForState(theme.useMaterial3
             ? _m3InputStyle(context)
             : theme.textTheme.titleMedium!)
-        .merge(widget.style);
+        .merge(providedStyle);
     final Brightness keyboardAppearance =
         widget.keyboardAppearance ?? theme.brightness;
     final TextEditingController controller = _effectiveController;
@@ -496,6 +498,7 @@ class ExtendedTextFieldState extends _TextFieldState {
           onEditingComplete: widget.onEditingComplete,
           onSubmitted: widget.onSubmitted,
           onAppPrivateCommand: widget.onAppPrivateCommand,
+          groupId: widget.groupId,
           onSelectionHandleTapped: _handleSelectionHandleTapped,
           onTapOutside: widget.onTapOutside,
           inputFormatters: formatters,
@@ -533,7 +536,6 @@ class ExtendedTextFieldState extends _TextFieldState {
               TextMagnifier.adaptiveMagnifierConfiguration,
           // zmtzawqlp
           specialTextSpanBuilder: extenedTextField.specialTextSpanBuilder,
-          onPasteFunction: extenedTextField.onPasteFunction,
         ),
       ),
     );
@@ -560,7 +562,7 @@ class ExtendedTextFieldState extends _TextFieldState {
     final MouseCursor effectiveMouseCursor =
         MaterialStateProperty.resolveAs<MouseCursor>(
       widget.mouseCursor ?? MaterialStateMouseCursor.textable,
-      _materialState,
+      _statesController.value,
     );
 
     final int? semanticsMaxValueLength;
@@ -578,11 +580,12 @@ class ExtendedTextFieldState extends _TextFieldState {
       onExit: (PointerExitEvent event) => _handleHover(false),
       child: TextFieldTapRegion(
         child: IgnorePointer(
-          ignoring: !_isEnabled,
+          ignoring: widget.ignorePointers ?? !_isEnabled,
           child: AnimatedBuilder(
             animation: controller, // changes the _currentLength
             builder: (BuildContext context, Widget? child) {
               return Semantics(
+                enabled: _isEnabled,
                 maxValueLength: semanticsMaxValueLength,
                 currentValueLength: _currentLength,
                 onTap: widget.readOnly
@@ -597,6 +600,35 @@ class ExtendedTextFieldState extends _TextFieldState {
                       },
                 onDidGainAccessibilityFocus: handleDidGainAccessibilityFocus,
                 onDidLoseAccessibilityFocus: handleDidLoseAccessibilityFocus,
+                onFocus: _isEnabled
+                    ? () {
+                        assert(
+                            _effectiveFocusNode.canRequestFocus,
+                            'Received SemanticsAction.focus from the engine. However, the FocusNode '
+                            'of this text field cannot gain focus. This likely indicates a bug. '
+                            'If this text field cannot be focused (e.g. because it is not '
+                            'enabled), then its corresponding semantics node must be configured '
+                            'such that the assistive technology cannot request focus on it.');
+
+                        if (_effectiveFocusNode.canRequestFocus &&
+                            !_effectiveFocusNode.hasFocus) {
+                          _effectiveFocusNode.requestFocus();
+                        } else if (!widget.readOnly) {
+                          // If the platform requested focus, that means that previously the
+                          // platform believed that the text field did not have focus (even
+                          // though Flutter's widget system believed otherwise). This likely
+                          // means that the on-screen keyboard is hidden, or more generally,
+                          // there is no current editing session in this field. To correct
+                          // that, keyboard must be requested.
+                          //
+                          // A concrete scenario where this can happen is when the user
+                          // dismisses the keyboard on the web. The editing session is
+                          // closed by the engine, but the text field widget stays focused
+                          // in the framework.
+                          _requestKeyboard();
+                        }
+                      }
+                    : null,
                 child: child,
               );
             },
