@@ -232,9 +232,10 @@ class _DiscreteKeyFrameSimulation extends Simulation {
 /// This widget provides default [Action]s for handling common text editing
 /// [Intent]s such as deleting, copying and pasting in the text field. These
 /// [Action]s can be directly invoked using [Actions.invoke] or the
-/// [Actions.maybeInvoke] method. The default text editing keyboard [Shortcuts]
-/// also use these [Intent]s and [Action]s to perform the text editing
-/// operations they are bound to.
+/// [Actions.maybeInvoke] method. The default text editing keyboard [Shortcuts],
+/// typically declared in [DefaultTextEditingShortcuts], also use these
+/// [Intent]s and [Action]s to perform the text editing operations they are
+/// bound to.
 ///
 /// The default handling of a specific [Intent] can be overridden by placing an
 /// [Actions] widget above this widget. See the [Action] class and the
@@ -284,6 +285,40 @@ class _DiscreteKeyFrameSimulation extends Simulation {
 /// | [CopySelectionTextIntent]               | Copies or cuts the selected text into the clipboard |
 /// | [PasteTextIntent]                       | Inserts the current text in the clipboard after the caret location, or replaces the selected text if the selection is not collapsed. |
 ///
+/// ## Text Editing [Shortcuts]
+///
+/// It's also possible to directly remap keyboard shortcuts to new [Intent]s by
+/// inserting a [Shortcuts] widget above this in the widget tree. When using
+/// [WidgetsApp], the large set of default text editing keyboard shortcuts are
+/// declared near the top of the widget tree in [DefaultTextEditingShortcuts],
+/// and any [Shortcuts] widget between it and this [EditableText] will override
+/// those defaults.
+///
+/// {@template flutter.widgets.editableText.shortcutsAndTextInput}
+/// ### Interactions Between [Shortcuts] and Text Input
+///
+/// Shortcuts prevent text input fields from receiving their keystrokes as text
+/// input. For example, placing a [Shortcuts] widget in the widget tree above
+/// a text input field and creating a shortcut for [LogicalKeyboardKey.keyA]
+/// will prevent the field from receiving that key as text input. In other
+/// words, typing key "A" into the field will trigger the shortcut and will not
+/// insert a letter "a" into the field.
+///
+/// This happens because of the way that key strokes are handled in Flutter.
+/// When a keystroke is received in Flutter's engine, it first gives the
+/// framework the opportunity to handle it as a raw key event through
+/// [SystemChannels.keyEvent]. This is what [Shortcuts] listens to indirectly
+/// through its [FocusNode]. If it is not handled, then it will proceed to try
+/// handling it as text input through [SystemChannels.textInput], which is what
+/// [EditableTextState] listens to through [TextInputClient].
+///
+/// This behavior, where a shortcut prevents text input into some field, can be
+/// overridden by using another [Shortcuts] widget lower in the widget tree and
+/// mapping the desired key stroke(s) to [DoNothingAndStopPropagationIntent].
+/// The key event will be reported as unhandled by the framework and will then
+/// be sent as text input as usual.
+/// {@endtemplate}
+///
 /// ## Gesture Events Handling
 ///
 /// When [rendererIgnoresPointer] is false (the default), this widget provides
@@ -306,6 +341,14 @@ class _DiscreteKeyFrameSimulation extends Simulation {
 ///    text when the text field is not [readOnly].
 ///  * When the virtual keyboard pops up.
 /// {@endtemplate}
+///
+/// ## Scrolling Considerations
+///
+/// If this [EditableText] is not a descendant of [Scaffold] and is being used
+/// within a [Scrollable] or nested [Scrollable]s, consider placing a
+/// [ScrollNotificationObserver] above the root [Scrollable] that contains this
+/// [EditableText] to ensure proper scroll coordination for [EditableText] and
+/// its components like [TextSelectionOverlay].
 ///
 /// {@template flutter.widgets.editableText.accessibility}
 /// ## Troubleshooting Common Accessibility Issues
@@ -390,6 +433,7 @@ class _EditableText extends StatefulWidget {
     this.onAppPrivateCommand,
     this.onSelectionChanged,
     this.onSelectionHandleTapped,
+    this.groupId = EditableText,
     this.onTapOutside,
     List<TextInputFormatter>? inputFormatters,
     this.mouseCursor,
@@ -474,7 +518,8 @@ class _EditableText extends StatefulWidget {
         assert(
           spellCheckConfiguration == null ||
               spellCheckConfiguration ==
-                  const SpellCheckConfiguration.disabled() ||
+                  // zmtzawqlp
+                  const _SpellCheckConfiguration.disabled() ||
               spellCheckConfiguration.misspelledTextStyle != null,
           'spellCheckConfiguration must specify a misspelledTextStyle if spell check behavior is desired',
         ),
@@ -628,7 +673,7 @@ class _EditableText extends StatefulWidget {
     if (_strutStyle == null) {
       return StrutStyle.fromTextStyle(style, forceStrutHeight: true);
     }
-    return _strutStyle!.inheritFromTextStyle(style);
+    return _strutStyle.inheritFromTextStyle(style);
   }
 
   final StrutStyle? _strutStyle;
@@ -714,6 +759,11 @@ class _EditableText extends StatefulWidget {
   /// while rendering the floating cursor.
   ///
   /// Typically this would be set to [CupertinoColors.inactiveGray].
+  ///
+  /// See also:
+  ///
+  ///  * [FloatingCursorDragState], which explains the floating cursor feature
+  ///    in detail.
   final Color backgroundCursorColor;
 
   /// {@template flutter.widgets.editableText.maxLines}
@@ -1016,6 +1066,19 @@ class _EditableText extends StatefulWidget {
 
   /// {@macro flutter.widgets.SelectionOverlay.onSelectionHandleTapped}
   final VoidCallback? onSelectionHandleTapped;
+
+  /// {@template flutter.widgets.editableText.groupId}
+  /// The group identifier for the [TextFieldTapRegion] of this text field.
+  ///
+  /// Text fields with the same group identifier share the same tap region.
+  /// Defaults to the type of [EditableText].
+  ///
+  /// See also:
+  ///
+  ///  * [TextFieldTapRegion], to give a [groupId] to a widget that is to be
+  ///    included in a [EditableText]'s tap region that has [groupId] set.
+  /// {@endtemplate}
+  final Object groupId;
 
   /// {@template flutter.widgets.editableText.onTapOutside}
   /// Called for each tap that occurs outside of the[TextFieldTapRegion] group
@@ -1403,11 +1466,28 @@ class _EditableText extends StatefulWidget {
   /// {@template flutter.widgets.EditableText.contextMenuBuilder}
   /// Builds the text selection toolbar when requested by the user.
   ///
-  /// `primaryAnchor` is the desired anchor position for the context menu, while
-  /// `secondaryAnchor` is the fallback location if the menu doesn't fit.
+  /// The context menu is built when [EditableTextState.showToolbar] is called,
+  /// typically by one of the callbacks installed by the widget created by
+  /// [TextSelectionGestureDetectorBuilder.buildGestureDetector]. The widget
+  /// returned by [contextMenuBuilder] is passed to a [ContextMenuController].
   ///
-  /// `buttonItems` represents the buttons that would be built by default for
-  /// this widget.
+  /// If no callback is provided, no context menu will be shown.
+  ///
+  /// The [EditableTextContextMenuBuilder] signature used by the
+  /// [contextMenuBuilder] callback has two parameters, the [BuildContext] of
+  /// the [EditableText] and the [EditableTextState] of the [EditableText].
+  ///
+  /// The [EditableTextState] has two properties that are especially useful when
+  /// building the widgets for the context menu:
+  ///
+  /// * [EditableTextState.contextMenuAnchors] specifies the desired anchor
+  ///   position for the context menu.
+  ///
+  /// * [EditableTextState.contextMenuButtonItems] represents the buttons that
+  ///   should typically be built for this widget (e.g. cut, copy, paste).
+  ///
+  /// The [TextSelectionToolbarLayoutDelegate] class may be particularly useful
+  /// in honoring the preferred anchor positions.
   ///
   /// For backwards compatibility, when [selectionControls] is set to an object
   /// that does not mix in [TextSelectionHandleControls], [contextMenuBuilder]
@@ -1438,8 +1518,6 @@ class _EditableText extends StatefulWidget {
   ///   * [BrowserContextMenu], which allows the browser's context menu on web
   ///     to be disabled and Flutter-rendered context menus to appear.
   /// {@endtemplate}
-  ///
-  /// If not provided, no context menu will be shown.
   final EditableTextContextMenuBuilder? contextMenuBuilder;
 
   /// {@template flutter.widgets.EditableText.spellCheckConfiguration}
@@ -1457,11 +1535,10 @@ class _EditableText extends StatefulWidget {
   // zmtzawqlp
   final _SpellCheckConfiguration? spellCheckConfiguration;
 
-  /// {@macro flutter.widgets.magnifier.TextMagnifierConfiguration.intro}
+  /// The configuration for the magnifier to use with selections in this text
+  /// field.
   ///
   /// {@macro flutter.widgets.magnifier.intro}
-  ///
-  /// {@macro flutter.widgets.magnifier.TextMagnifierConfiguration.details}
   final TextMagnifierConfiguration magnifierConfiguration;
 
   bool get _userSelectionEnabled =>
@@ -1503,6 +1580,11 @@ class _EditableText extends StatefulWidget {
       // If the paste button is enabled, don't render anything until the state
       // of the clipboard is known, since it's used to determine if paste is
       // shown.
+
+      // On Android, the share button is before the select all button.
+      final bool showShareBeforeSelectAll =
+          defaultTargetPlatform == TargetPlatform.android;
+
       resultButtonItem.addAll(<ContextMenuButtonItem>[
         if (onCut != null)
           ContextMenuButtonItem(
@@ -1519,6 +1601,11 @@ class _EditableText extends StatefulWidget {
             onPressed: onPaste,
             type: ContextMenuButtonType.paste,
           ),
+        if (onShare != null && showShareBeforeSelectAll)
+          ContextMenuButtonItem(
+            onPressed: onShare,
+            type: ContextMenuButtonType.share,
+          ),
         if (onSelectAll != null)
           ContextMenuButtonItem(
             onPressed: onSelectAll,
@@ -1534,7 +1621,7 @@ class _EditableText extends StatefulWidget {
             onPressed: onSearchWeb,
             type: ContextMenuButtonType.searchWeb,
           ),
-        if (onShare != null)
+        if (onShare != null && !showShareBeforeSelectAll)
           ContextMenuButtonItem(
             onPressed: onShare,
             type: ContextMenuButtonType.share,
@@ -1831,6 +1918,14 @@ class _EditableTextState extends State<_EditableText>
 
   /// zmtzawqlp
   _TextSelectionOverlay? _selectionOverlay;
+  ScrollNotificationObserverState? _scrollNotificationObserver;
+  ({
+    TextEditingValue value,
+    Rect selectionBounds
+  })? _dataWhenToolbarShowScheduled;
+  bool _listeningToScrollNotificationObserver = false;
+
+  bool get _webContextMenuEnabled => kIsWeb && BrowserContextMenu.enabled;
 
   final GlobalKey _scrollableKey = GlobalKey();
   ScrollController? _internalScrollController;
@@ -1869,7 +1964,7 @@ class _EditableTextState extends State<_EditableText>
 
   /// Whether or not spell check is enabled.
   ///
-  /// Spell check is enabled when a [_SpellCheckConfiguration] has been specified
+  /// Spell check is enabled when a [SpellCheckConfiguration] has been specified
   /// for the widget.
   bool get spellCheckEnabled => _spellCheckConfiguration.spellCheckEnabled;
 
@@ -1884,6 +1979,12 @@ class _EditableTextState extends State<_EditableText>
       spellCheckEnabled &&
       spellCheckResults != null &&
       spellCheckResults!.suggestionSpans.isNotEmpty;
+
+  /// The text processing service used to retrieve the native text processing actions.
+  final ProcessTextService _processTextService = DefaultProcessTextService();
+
+  /// The list of native text processing actions provided by the engine.
+  final List<ProcessTextAction> _processTextActions = <ProcessTextAction>[];
 
   /// Whether to create an input connection with the platform for text editing
   /// or not.
@@ -1996,14 +2097,21 @@ class _EditableTextState extends State<_EditableText>
 
   @override
   bool get shareEnabled {
-    if (defaultTargetPlatform != TargetPlatform.iOS) {
-      return false;
+    switch (defaultTargetPlatform) {
+      case TargetPlatform.android:
+      case TargetPlatform.iOS:
+        return !widget.obscureText &&
+            !textEditingValue.selection.isCollapsed &&
+            textEditingValue.selection
+                    .textInside(textEditingValue.text)
+                    .trim() !=
+                '';
+      case TargetPlatform.macOS:
+      case TargetPlatform.fuchsia:
+      case TargetPlatform.linux:
+      case TargetPlatform.windows:
+        return false;
     }
-
-    return !widget.obscureText &&
-        !textEditingValue.selection.isCollapsed &&
-        textEditingValue.selection.textInside(textEditingValue.text).trim() !=
-            '';
   }
 
   @override
@@ -2088,20 +2196,20 @@ class _EditableTextState extends State<_EditableText>
         if (mounted) {
           bringIntoView(textEditingValue.selection.extent);
         }
-      });
+      }, debugLabel: 'EditableText.bringSelectionIntoView');
       hideToolbar();
     }
     clipboardStatus.update();
   }
 
+  bool get _allowPaste {
+    return !widget.readOnly && textEditingValue.selection.isValid;
+  }
+
   /// Paste text from [Clipboard].
   @override
   Future<void> pasteText(SelectionChangedCause cause) async {
-    if (widget.readOnly) {
-      return;
-    }
-    final TextSelection selection = textEditingValue.selection;
-    if (!selection.isValid) {
+    if (!_allowPaste) {
       return;
     }
     // Snapshot the input before using `await`.
@@ -2110,9 +2218,17 @@ class _EditableTextState extends State<_EditableText>
     if (data == null) {
       return;
     }
+    _pasteText(cause, data.text!);
+  }
+
+  void _pasteText(SelectionChangedCause cause, String text) {
+    if (!_allowPaste) {
+      return;
+    }
 
     // After the paste, the cursor should be collapsed and located after the
     // pasted content.
+    final TextSelection selection = textEditingValue.selection;
     final int lastSelectionIndex =
         math.max(selection.baseOffset, selection.extentOffset);
     final TextEditingValue collapsedTextEditingValue =
@@ -2121,7 +2237,7 @@ class _EditableTextState extends State<_EditableText>
     );
 
     userUpdateTextEditingValue(
-      collapsedTextEditingValue.replaced(selection, data.text!),
+      collapsedTextEditingValue.replaced(selection, text),
       cause,
     );
     if (cause == SelectionChangedCause.toolbar) {
@@ -2130,7 +2246,7 @@ class _EditableTextState extends State<_EditableText>
         if (mounted) {
           bringIntoView(textEditingValue.selection.extent);
         }
-      });
+      }, debugLabel: 'EditableText.bringSelectionIntoView');
       hideToolbar();
     }
   }
@@ -2219,9 +2335,9 @@ class _EditableTextState extends State<_EditableText>
   }
 
   /// Launch the share interface for the current selection,
-  /// as in the "Share" edit menu button on iOS.
+  /// as in the "Share..." edit menu button on iOS.
   ///
-  /// Currently this is only implemented for iOS.
+  /// Currently this is only implemented for iOS and Android.
   ///
   /// When 'obscureText' is true or the selection is empty,
   /// this function will not do anything
@@ -2381,7 +2497,12 @@ class _EditableTextState extends State<_EditableText>
 
   /// Gets the line heights at the start and end of the selection for the given
   /// [_EditableTextState].
-  _GlyphHeights _getGlyphHeights() {
+  ///
+  /// See also:
+  ///
+  /// * [TextSelectionToolbarAnchors.getSelectionRect], which depends on this
+  ///   information.
+  ({double startGlyphHeight, double endGlyphHeight}) getGlyphHeights() {
     final TextSelection selection = textEditingValue.selection;
 
     // Only calculate handle rects if the text in the previous frame
@@ -2395,9 +2516,9 @@ class _EditableTextState extends State<_EditableText>
     final String prevText = span.toPlainText();
     final String currText = textEditingValue.text;
     if (prevText != currText || !selection.isValid || selection.isCollapsed) {
-      return _GlyphHeights(
-        start: renderEditable.preferredLineHeight,
-        end: renderEditable.preferredLineHeight,
+      return (
+        startGlyphHeight: renderEditable.preferredLineHeight,
+        endGlyphHeight: renderEditable.preferredLineHeight,
       );
     }
 
@@ -2416,9 +2537,11 @@ class _EditableTextState extends State<_EditableText>
       start: selection.end - lastSelectedGraphemeExtent,
       end: selection.end,
     ));
-    return _GlyphHeights(
-      start: startCharacterRect?.height ?? renderEditable.preferredLineHeight,
-      end: endCharacterRect?.height ?? renderEditable.preferredLineHeight,
+    return (
+      startGlyphHeight:
+          startCharacterRect?.height ?? renderEditable.preferredLineHeight,
+      endGlyphHeight:
+          endCharacterRect?.height ?? renderEditable.preferredLineHeight,
     );
   }
 
@@ -2437,14 +2560,17 @@ class _EditableTextState extends State<_EditableText>
       );
     }
 
-    final _GlyphHeights glyphHeights = _getGlyphHeights();
+    final (
+      startGlyphHeight: double startGlyphHeight,
+      endGlyphHeight: double endGlyphHeight
+    ) = getGlyphHeights();
     final TextSelection selection = textEditingValue.selection;
     final List<TextSelectionPoint> points =
         renderEditable.getEndpointsForSelection(selection);
     return TextSelectionToolbarAnchors.fromSelection(
       renderBox: renderEditable,
-      startGlyphHeight: glyphHeights.start,
-      endGlyphHeight: glyphHeights.end,
+      startGlyphHeight: startGlyphHeight,
+      endGlyphHeight: endGlyphHeight,
       selectionEndpoints: points,
     );
   }
@@ -2493,7 +2619,38 @@ class _EditableTextState extends State<_EditableText>
           onLiveTextInput: liveTextInputEnabled
               ? () => _startLiveTextInput(SelectionChangedCause.toolbar)
               : null,
-        );
+        )
+      ..addAll(_textProcessingActionButtonItems);
+  }
+
+  List<ContextMenuButtonItem> get _textProcessingActionButtonItems {
+    final List<ContextMenuButtonItem> buttonItems = <ContextMenuButtonItem>[];
+    final TextSelection selection = textEditingValue.selection;
+    if (widget.obscureText || !selection.isValid || selection.isCollapsed) {
+      return buttonItems;
+    }
+
+    for (final ProcessTextAction action in _processTextActions) {
+      buttonItems.add(ContextMenuButtonItem(
+        label: action.label,
+        onPressed: () async {
+          final String selectedText =
+              selection.textInside(textEditingValue.text);
+          if (selectedText.isNotEmpty) {
+            final String? processedText = await _processTextService
+                .processTextAction(action.id, selectedText, widget.readOnly);
+            // If an activity does not return a modified version, just hide the toolbar.
+            // Otherwise use the result to replace the selected text.
+            if (processedText != null && _allowPaste) {
+              _pasteText(SelectionChangedCause.toolbar, processedText);
+            } else {
+              hideToolbar();
+            }
+          }
+        },
+      ));
+    }
+    return buttonItems;
   }
 
   // State lifecycle:
@@ -2505,11 +2662,18 @@ class _EditableTextState extends State<_EditableText>
     clipboardStatus.addListener(_onChangedClipboardStatus);
     widget.controller.addListener(_didChangeTextEditingValue);
     widget.focusNode.addListener(_handleFocusChanged);
-    _scrollController.addListener(_onEditableScroll);
     _cursorVisibilityNotifier.value = widget.showCursor;
     // zmtzawqlp
     // _spellCheckConfiguration =
     //     _inferSpellCheckConfiguration(widget.spellCheckConfiguration);
+    _initProcessTextActions();
+  }
+
+  /// Query the engine to initialize the list of text processing actions to show
+  /// in the text selection toolbar.
+  Future<void> _initProcessTextActions() async {
+    _processTextActions.clear();
+    _processTextActions.addAll(await _processTextService.queryTextActions());
   }
 
   // Whether `TickerMode.of(context)` is true and animations (like blinking the
@@ -2538,7 +2702,7 @@ class _EditableTextState extends State<_EditableText>
           _flagInternalFocus();
           FocusScope.of(context).autofocus(widget.focusNode);
         }
-      });
+      }, debugLabel: 'EditableText.autofocus');
     }
 
     // Restart or stop the blinking cursor when TickerMode changes.
@@ -2549,6 +2713,15 @@ class _EditableTextState extends State<_EditableText>
         _startCursorBlink();
       } else if (!_tickersEnabled && _cursorTimer != null) {
         _stopCursorBlink();
+      }
+    }
+
+    // Check for changes in viewId.
+    if (_hasInputConnection) {
+      final int newViewId = View.of(context).viewId;
+      if (newViewId != _viewId) {
+        _textInputConnection!
+            .updateConfig(_effectiveAutofillClient.textInputConfiguration);
       }
     }
 
@@ -2572,6 +2745,18 @@ class _EditableTextState extends State<_EditableText>
         hideToolbar();
       }
     }
+
+    if (_listeningToScrollNotificationObserver) {
+      // Only update subscription when we have previously subscribed to the
+      // scroll notification observer. We only subscribe to the scroll
+      // notification observer when the context menu is shown on platforms that
+      // support _platformSupportsFadeOnScroll.
+      _scrollNotificationObserver
+          ?.removeListener(_handleContextMenuOnParentScroll);
+      _scrollNotificationObserver = ScrollNotificationObserver.maybeOf(context);
+      _scrollNotificationObserver
+          ?.addListener(_handleContextMenuOnParentScroll);
+    }
   }
 
   @override
@@ -2584,7 +2769,30 @@ class _EditableTextState extends State<_EditableText>
       widget.controller.addListener(_didChangeTextEditingValue);
       _updateRemoteEditingValueIfNeeded();
     }
-    if (widget.controller.selection != oldWidget.controller.selection) {
+
+    if (_selectionOverlay != null &&
+        (widget.contextMenuBuilder != oldWidget.contextMenuBuilder ||
+            widget.selectionControls != oldWidget.selectionControls ||
+            widget.onSelectionHandleTapped !=
+                oldWidget.onSelectionHandleTapped ||
+            widget.dragStartBehavior != oldWidget.dragStartBehavior ||
+            widget.magnifierConfiguration !=
+                oldWidget.magnifierConfiguration)) {
+      final bool shouldShowToolbar = _selectionOverlay!.toolbarIsVisible;
+      final bool shouldShowHandles = _selectionOverlay!.handlesVisible;
+      _selectionOverlay!.dispose();
+      _selectionOverlay = _createSelectionOverlay();
+      if (shouldShowToolbar || shouldShowHandles) {
+        SchedulerBinding.instance.addPostFrameCallback((Duration _) {
+          if (shouldShowToolbar) {
+            _selectionOverlay!.showToolbar();
+          }
+          if (shouldShowHandles) {
+            _selectionOverlay!.showHandles();
+          }
+        });
+      }
+    } else if (widget.controller.selection != oldWidget.controller.selection) {
       _selectionOverlay?.update(_value);
     }
     _selectionOverlay?.handlesVisible = widget.showSelectionHandles;
@@ -2601,12 +2809,6 @@ class _EditableTextState extends State<_EditableText>
       updateKeepAlive();
     }
 
-    if (widget.scrollController != oldWidget.scrollController) {
-      (oldWidget.scrollController ?? _internalScrollController)
-          ?.removeListener(_onEditableScroll);
-      _scrollController.addListener(_onEditableScroll);
-    }
-
     if (!_shouldCreateInputConnection) {
       _closeInputConnectionIfNeeded();
     } else if (oldWidget.readOnly && _hasFocus) {
@@ -2614,7 +2816,7 @@ class _EditableTextState extends State<_EditableText>
       // See https://github.com/flutter/flutter/issues/126312
       SchedulerBinding.instance.addPostFrameCallback((Duration _) {
         _openInputConnection();
-      });
+      }, debugLabel: 'EditableText.openInputConnection');
     }
 
     if (kIsWeb && _hasInputConnection) {
@@ -2660,6 +2862,15 @@ class _EditableTextState extends State<_EditableText>
     }
   }
 
+  void _disposeScrollNotificationObserver() {
+    _listeningToScrollNotificationObserver = false;
+    if (_scrollNotificationObserver != null) {
+      _scrollNotificationObserver!
+          .removeListener(_handleContextMenuOnParentScroll);
+      _scrollNotificationObserver = null;
+    }
+  }
+
   @override
   void dispose() {
     _internalScrollController?.dispose();
@@ -2683,6 +2894,7 @@ class _EditableTextState extends State<_EditableText>
     clipboardStatus.dispose();
     _cursorVisibilityNotifier.dispose();
     FocusManager.instance.removeListener(_unflagInternalFocus);
+    _disposeScrollNotificationObserver();
     super.dispose();
     assert(_batchEditDepth <= 0, 'unfinished batch edits: $_batchEditDepth');
   }
@@ -2832,7 +3044,7 @@ class _EditableTextState extends State<_EditableText>
   }
 
   // The original position of the caret on FloatingCursorDragState.start.
-  Rect? _startCaretRect;
+  Offset? _startCaretCenter;
 
   // The most recent text position as determined by the location of the floating
   // cursor.
@@ -2868,20 +3080,34 @@ class _EditableTextState extends State<_EditableText>
         // we cache the position.
         _pointOffsetOrigin = point.offset;
 
-        final TextPosition currentTextPosition = TextPosition(
-            offset: renderEditable.selection!.baseOffset,
-            affinity: renderEditable.selection!.affinity);
-        _startCaretRect =
-            renderEditable.getLocalRectForCaret(currentTextPosition);
+        final Offset startCaretCenter;
+        final TextPosition currentTextPosition;
+        final bool shouldResetOrigin;
+        // Only non-null when starting a floating cursor via long press.
+        if (point.startLocation != null) {
+          shouldResetOrigin = false;
+          (startCaretCenter, currentTextPosition) = point.startLocation!;
+        } else {
+          shouldResetOrigin = true;
+          currentTextPosition = TextPosition(
+              offset: renderEditable.selection!.baseOffset,
+              affinity: renderEditable.selection!.affinity);
+          startCaretCenter =
+              renderEditable.getLocalRectForCaret(currentTextPosition).center;
+        }
 
-        _lastBoundedOffset = _startCaretRect!.center - _floatingCursorOffset;
+        _startCaretCenter = startCaretCenter;
+        _lastBoundedOffset =
+            renderEditable.calculateBoundedFloatingCursorOffset(
+                _startCaretCenter! - _floatingCursorOffset,
+                shouldResetOrigin: shouldResetOrigin);
         _lastTextPosition = currentTextPosition;
         renderEditable.setFloatingCursor(
             point.state, _lastBoundedOffset!, _lastTextPosition!);
       case FloatingCursorDragState.Update:
         final Offset centeredPoint = point.offset! - _pointOffsetOrigin!;
         final Offset rawCursorOffset =
-            _startCaretRect!.center + centeredPoint - _floatingCursorOffset;
+            _startCaretCenter! + centeredPoint - _floatingCursorOffset;
 
         _lastBoundedOffset = renderEditable
             .calculateBoundedFloatingCursorOffset(rawCursorOffset);
@@ -2929,7 +3155,7 @@ class _EditableTextState extends State<_EditableText>
         _handleSelectionChanged(TextSelection.fromPosition(_lastTextPosition!),
             SelectionChangedCause.forcePress);
       }
-      _startCaretRect = null;
+      _startCaretCenter = null;
       _lastTextPosition = null;
       _pointOffsetOrigin = null;
       _lastBoundedOffset = null;
@@ -3298,12 +3524,180 @@ class _EditableTextState extends State<_EditableText>
     }
   }
 
-  void _onEditableScroll() {
-    _selectionOverlay?.updateForScroll();
-    _scribbleCacheKey = null;
+  final bool _platformSupportsFadeOnScroll = switch (defaultTargetPlatform) {
+    TargetPlatform.android || TargetPlatform.iOS => true,
+    TargetPlatform.fuchsia ||
+    TargetPlatform.linux ||
+    TargetPlatform.macOS ||
+    TargetPlatform.windows =>
+      false,
+  };
+
+  bool _isInternalScrollableNotification(BuildContext? notificationContext) {
+    final ScrollableState? scrollableState =
+        notificationContext?.findAncestorStateOfType<ScrollableState>();
+    return _scrollableKey.currentContext == scrollableState?.context;
   }
 
-  /// zmtzawqlp
+  bool _scrollableNotificationIsFromSameSubtree(
+      BuildContext? notificationContext) {
+    if (notificationContext == null) {
+      return false;
+    }
+    BuildContext? currentContext = context;
+    // The notification context of a ScrollNotification points to the RawGestureDetector
+    // of the Scrollable. We get the ScrollableState associated with this notification
+    // by looking up the tree.
+    final ScrollableState? notificationScrollableState =
+        notificationContext.findAncestorStateOfType<ScrollableState>();
+    if (notificationScrollableState == null) {
+      return false;
+    }
+    while (currentContext != null) {
+      final ScrollableState? scrollableState =
+          currentContext.findAncestorStateOfType<ScrollableState>();
+      if (scrollableState == notificationScrollableState) {
+        return true;
+      }
+      currentContext = scrollableState?.context;
+    }
+    return false;
+  }
+
+  void _handleContextMenuOnParentScroll(ScrollNotification notification) {
+    // Do some preliminary checks to avoid expensive subtree traversal.
+    if (notification is! ScrollStartNotification &&
+        notification is! ScrollEndNotification) {
+      return;
+    }
+    switch (notification) {
+      case ScrollStartNotification() when _dataWhenToolbarShowScheduled != null:
+      case ScrollEndNotification() when _dataWhenToolbarShowScheduled == null:
+        break;
+      case ScrollEndNotification()
+          when _dataWhenToolbarShowScheduled!.value != _value:
+        _dataWhenToolbarShowScheduled = null;
+        _disposeScrollNotificationObserver();
+      case ScrollNotification(:final BuildContext? context)
+          when !_isInternalScrollableNotification(context) &&
+              _scrollableNotificationIsFromSameSubtree(context):
+        _handleContextMenuOnScroll(notification);
+    }
+  }
+
+  Rect _calculateDeviceRect() {
+    final Size screenSize = MediaQuery.sizeOf(context);
+    final ui.FlutterView view = View.of(context);
+    final double obscuredVertical =
+        (view.padding.top + view.padding.bottom + view.viewInsets.bottom) /
+            view.devicePixelRatio;
+    final double obscuredHorizontal =
+        (view.padding.left + view.padding.right) / view.devicePixelRatio;
+    final Size visibleScreenSize = Size(screenSize.width - obscuredHorizontal,
+        screenSize.height - obscuredVertical);
+    return Rect.fromLTWH(
+        view.padding.left / view.devicePixelRatio,
+        view.padding.top / view.devicePixelRatio,
+        visibleScreenSize.width,
+        visibleScreenSize.height);
+  }
+
+  bool _showToolbarOnScreenScheduled = false;
+  void _handleContextMenuOnScroll(ScrollNotification notification) {
+    if (_webContextMenuEnabled) {
+      return;
+    }
+    if (!_platformSupportsFadeOnScroll) {
+      _selectionOverlay?.updateForScroll();
+      return;
+    }
+    // When the scroll begins and the toolbar is visible, hide it
+    // until scrolling ends.
+    //
+    // The selection and renderEditable need to be visible within the current
+    // viewport for the toolbar to show when scrolling ends. If they are not
+    // then the toolbar is shown when they are scrolled back into view, unless
+    // invalidated by a change in TextEditingValue.
+    if (notification is ScrollStartNotification) {
+      if (_dataWhenToolbarShowScheduled != null) {
+        return;
+      }
+      final bool toolbarIsVisible = _selectionOverlay != null &&
+          _selectionOverlay!.toolbarIsVisible &&
+          !_selectionOverlay!.spellCheckToolbarIsVisible;
+      if (!toolbarIsVisible) {
+        return;
+      }
+      final List<TextBox> selectionBoxes =
+          renderEditable.getBoxesForSelection(_value.selection);
+      final Rect selectionBounds = _value.selection.isCollapsed ||
+              selectionBoxes.isEmpty
+          ? renderEditable.getLocalRectForCaret(_value.selection.extent)
+          : selectionBoxes
+              .map((TextBox box) => box.toRect())
+              .reduce((Rect result, Rect rect) => result.expandToInclude(rect));
+      _dataWhenToolbarShowScheduled =
+          (value: _value, selectionBounds: selectionBounds);
+      _selectionOverlay?.hideToolbar();
+    } else if (notification is ScrollEndNotification) {
+      if (_dataWhenToolbarShowScheduled == null) {
+        return;
+      }
+      if (_dataWhenToolbarShowScheduled!.value != _value) {
+        // Value has changed so we should invalidate any toolbar scheduling.
+        _dataWhenToolbarShowScheduled = null;
+        _disposeScrollNotificationObserver();
+        return;
+      }
+
+      if (_showToolbarOnScreenScheduled) {
+        return;
+      }
+      _showToolbarOnScreenScheduled = true;
+      SchedulerBinding.instance.addPostFrameCallback((Duration _) {
+        _showToolbarOnScreenScheduled = false;
+        if (!mounted) {
+          return;
+        }
+        final Rect deviceRect = _calculateDeviceRect();
+        final bool selectionVisibleInEditable =
+            renderEditable.selectionStartInViewport.value ||
+                renderEditable.selectionEndInViewport.value;
+        final Rect selectionBounds = MatrixUtils.transformRect(
+            renderEditable.getTransformTo(null),
+            _dataWhenToolbarShowScheduled!.selectionBounds);
+        final bool selectionOverlapsWithDeviceRect =
+            !selectionBounds.hasNaN && deviceRect.overlaps(selectionBounds);
+
+        if (selectionVisibleInEditable &&
+            selectionOverlapsWithDeviceRect &&
+            _selectionInViewport(
+                _dataWhenToolbarShowScheduled!.selectionBounds)) {
+          showToolbar();
+          _dataWhenToolbarShowScheduled = null;
+        }
+      }, debugLabel: 'EditableText.scheduleToolbar');
+    }
+  }
+
+  bool _selectionInViewport(Rect selectionBounds) {
+    RenderAbstractViewport? closestViewport =
+        RenderAbstractViewport.maybeOf(renderEditable);
+    while (closestViewport != null) {
+      final Rect selectionBoundsLocalToViewport = MatrixUtils.transformRect(
+          renderEditable.getTransformTo(closestViewport), selectionBounds);
+      if (selectionBoundsLocalToViewport.hasNaN ||
+          closestViewport.paintBounds.hasNaN ||
+          !closestViewport.paintBounds
+              .overlaps(selectionBoundsLocalToViewport)) {
+        return false;
+      }
+      closestViewport = RenderAbstractViewport.maybeOf(closestViewport.parent);
+    }
+    return true;
+  }
+
+  // zmtzawqlp
   _TextSelectionOverlay _createSelectionOverlay() {
     // final EditableTextContextMenuBuilder? contextMenuBuilder =
     //     widget.contextMenuBuilder;
@@ -3321,7 +3715,7 @@ class _EditableTextState extends State<_EditableText>
       dragStartBehavior: widget.dragStartBehavior,
       onSelectionHandleTapped: widget.onSelectionHandleTapped,
       // zmtzawqlp
-      // contextMenuBuilder: contextMenuBuilder == null
+      // contextMenuBuilder: contextMenuBuilder == null || _webContextMenuEnabled
       //     ? null
       //     : (BuildContext context) {
       //         return contextMenuBuilder(
@@ -3341,7 +3735,8 @@ class _EditableTextState extends State<_EditableText>
     // We return early if the selection is not valid. This can happen when the
     // text of [EditableText] is updated at the same time as the selection is
     // changed by a gesture event.
-    if (!widget.controller.isSelectionWithinTextBounds(selection)) {
+    final String text = widget.controller.value.text;
+    if (text.length < selection.end || text.length < selection.start) {
       return;
     }
 
@@ -3416,6 +3811,7 @@ class _EditableTextState extends State<_EditableText>
       _showCaretOnScreenScheduled = false;
       // Since we are in a post frame callback, check currentContext in case
       // RenderEditable has been disposed (in which case it will be null).
+      // zmtzawqlp
       final _RenderEditable? renderEditable =
           _editableKey.currentContext?.findRenderObject() as _RenderEditable?;
       if (renderEditable == null ||
@@ -3492,7 +3888,7 @@ class _EditableTextState extends State<_EditableText>
           rect: caretPadding.inflateRect(rectToReveal),
         );
       }
-    });
+    }, debugLabel: 'EditableText.showCaret');
   }
 
   late double _lastBottomViewInset;
@@ -3506,7 +3902,7 @@ class _EditableTextState extends State<_EditableText>
     if (_lastBottomViewInset != view.viewInsets.bottom) {
       SchedulerBinding.instance.addPostFrameCallback((Duration _) {
         _selectionOverlay?.updateForScroll();
-      });
+      }, debugLabel: 'EditableText.updateForScroll');
       if (_lastBottomViewInset < view.viewInsets.bottom) {
         // Because the metrics change signal from engine will come here every frame
         // (on both iOS and Android). So we don't need to show caret with animation.
@@ -3645,17 +4041,21 @@ class _EditableTextState extends State<_EditableText>
   }
 
   void _onCursorColorTick() {
+    final double effectiveOpacity = math.min(
+        widget.cursorColor.alpha / 255.0, _cursorBlinkOpacityController.value);
     renderEditable.cursorColor =
-        widget.cursorColor.withOpacity(_cursorBlinkOpacityController.value);
-    _cursorVisibilityNotifier.value =
-        widget.showCursor && _cursorBlinkOpacityController.value > 0;
+        widget.cursorColor.withOpacity(effectiveOpacity);
+    _cursorVisibilityNotifier.value = widget.showCursor &&
+        (EditableText.debugDeterministicCursor ||
+            _cursorBlinkOpacityController.value > 0);
   }
 
   bool get _showBlinkingCursor =>
       _hasFocus &&
       _value.selection.isCollapsed &&
       widget.showCursor &&
-      _tickersEnabled;
+      _tickersEnabled &&
+      !renderEditable.floatingCursorOn;
 
   /// Whether the blinking cursor is actually visible at this precise moment
   /// (it's hidden half the time, since it blinks).
@@ -3734,7 +4134,10 @@ class _EditableTextState extends State<_EditableText>
   }
 
   void _stopCursorBlink({bool resetCharTicks = true}) {
-    _cursorBlinkOpacityController.value = 0.0;
+    // If the cursor is animating, stop the animation, and we always
+    // want the cursor to be visible when the floating cursor is enabled.
+    _cursorBlinkOpacityController.value =
+        renderEditable.floatingCursorOn ? 1.0 : 0.0;
     _cursorTimer?.cancel();
     _cursorTimer = null;
     if (resetCharTicks) {
@@ -3840,8 +4243,9 @@ class _EditableTextState extends State<_EditableText>
     _updateSelectionRects();
     _updateComposingRectIfNeeded();
     _updateCaretRectIfNeeded();
-    SchedulerBinding.instance
-        .addPostFrameCallback(_schedulePeriodicPostFrameCallbacks);
+    SchedulerBinding.instance.addPostFrameCallback(
+        _schedulePeriodicPostFrameCallbacks,
+        debugLabel: 'EditableText.postFrameCallbacks');
   }
 
   _ScribbleCacheKey? _scribbleCacheKey;
@@ -3859,11 +4263,21 @@ class _EditableTextState extends State<_EditableText>
     }
 
     final InlineSpan inlineSpan = renderEditable.text!;
+    double? textScaleFactor;
+    final TextScaler effectiveTextScaler = switch ((
+      widget.textScaler, textScaleFactor // widget.textScaleFactor
+    )) {
+      (final TextScaler textScaler, _) => textScaler,
+      (null, final double textScaleFactor) =>
+        TextScaler.linear(textScaleFactor),
+      (null, null) => MediaQuery.textScalerOf(context),
+    };
+
     final _ScribbleCacheKey newCacheKey = _ScribbleCacheKey(
       inlineSpan: inlineSpan,
       textAlign: widget.textAlign,
       textDirection: _textDirection,
-      textScaler: widget.textScaler ?? MediaQuery.textScalerOf(context),
+      textScaler: effectiveTextScaler,
       textHeightBehavior: widget.textHeightBehavior ??
           DefaultTextHeightBehavior.maybeOf(context),
       locale: widget.locale,
@@ -3917,11 +4331,13 @@ class _EditableTextState extends State<_EditableText>
     _textInputConnection!.setSelectionRects(rects);
   }
 
-  // Sends the current composing rect to the iOS text input plugin via the text
-  // input channel. We need to keep sending the information even if no text is
-  // currently marked, as the information usually lags behind. The text input
-  // plugin needs to estimate the composing rect based on the latest caret rect,
-  // when the composing rect info didn't arrive in time.
+  // Sends the current composing rect to the embedder's text input plugin.
+  //
+  // In cases where the composing rect hasn't been updated in the embedder due
+  // to the lag of asynchronous messages over the channel, the position of the
+  // current caret rect is used instead.
+  //
+  // See: [_updateCaretRectIfNeeded]
   void _updateComposingRectIfNeeded() {
     final TextRange composingRange = _value.composing;
     assert(mounted);
@@ -3929,7 +4345,6 @@ class _EditableTextState extends State<_EditableText>
         renderEditable.getRectForComposingRange(composingRange);
     // Send the caret location instead if there's no marked text yet.
     if (composingRect == null) {
-      assert(!composingRange.isValid || composingRange.isCollapsed);
       final int offset = composingRange.isValid ? composingRange.start : 0;
       composingRect =
           renderEditable.getLocalRectForCaret(TextPosition(offset: offset));
@@ -3937,13 +4352,25 @@ class _EditableTextState extends State<_EditableText>
     _textInputConnection!.setComposingRect(composingRect);
   }
 
+  // Sends the current caret rect to the embedder's text input plugin.
+  //
+  // The position of the caret rect is updated periodically such that if the
+  // user initiates composing input, the current cursor rect can be used for
+  // the first character until the composing rect can be sent.
+  //
+  // On selection changes, the start of the selection is used. This ensures
+  // that regardless of the direction the selection was created, the cursor is
+  // set to the position where next text input occurs. This position is used to
+  // position the IME's candidate selection menu.
+  //
+  // See: [_updateComposingRectIfNeeded]
   void _updateCaretRectIfNeeded() {
     final TextSelection? selection = renderEditable.selection;
-    if (selection == null || !selection.isValid || !selection.isCollapsed) {
+    if (selection == null || !selection.isValid) {
       return;
     }
     final TextPosition currentTextPosition =
-        TextPosition(offset: selection.baseOffset);
+        TextPosition(offset: selection.start);
     final Rect caretRect =
         renderEditable.getLocalRectForCaret(currentTextPosition);
     _textInputConnection!.setCaretRect(caretRect);
@@ -3983,7 +4410,7 @@ class _EditableTextState extends State<_EditableText>
       if (!widget.focusNode.hasFocus) {
         _flagInternalFocus();
         widget.focusNode.requestFocus();
-        _selectionOverlay = _createSelectionOverlay();
+        _selectionOverlay ??= _createSelectionOverlay();
       }
       return;
     }
@@ -4011,7 +4438,7 @@ class _EditableTextState extends State<_EditableText>
     // functionality depending on the browser (such as translate). Due to this,
     // we should not show a Flutter toolbar for the editable text elements
     // unless the browser's context menu is explicitly disabled.
-    if (kIsWeb && BrowserContextMenu.enabled) {
+    if (_webContextMenuEnabled) {
       return false;
     }
 
@@ -4021,11 +4448,23 @@ class _EditableTextState extends State<_EditableText>
     _liveTextInputStatus?.update();
     clipboardStatus.update();
     _selectionOverlay!.showToolbar();
+    // Listen to parent scroll events when the toolbar is visible so it can be
+    // hidden during a scroll on supported platforms.
+    if (_platformSupportsFadeOnScroll) {
+      _listeningToScrollNotificationObserver = true;
+      _scrollNotificationObserver
+          ?.removeListener(_handleContextMenuOnParentScroll);
+      _scrollNotificationObserver = ScrollNotificationObserver.maybeOf(context);
+      _scrollNotificationObserver
+          ?.addListener(_handleContextMenuOnParentScroll);
+    }
     return true;
   }
 
   @override
   void hideToolbar([bool hideHandles = true]) {
+    // Stop listening to parent scroll events when toolbar is hidden.
+    _disposeScrollNotificationObserver();
     if (hideHandles) {
       // Hide the handles and the toolbar.
       _selectionOverlay?.hide();
@@ -4039,7 +4478,6 @@ class _EditableTextState extends State<_EditableText>
   void toggleToolbar([bool hideHandles = true]) {
     final _TextSelectionOverlay selectionOverlay =
         _selectionOverlay ??= _createSelectionOverlay();
-
     if (selectionOverlay.toolbarIsVisible) {
       hideToolbar(hideHandles);
     } else {
@@ -4054,9 +4492,8 @@ class _EditableTextState extends State<_EditableText>
     // platforms. Additionally, the Cupertino style toolbar can't be drawn on
     // the web with the HTML renderer due to
     // https://github.com/flutter/flutter/issues/123560.
-    final bool platformNotSupported = kIsWeb && BrowserContextMenu.enabled;
     if (!spellCheckEnabled ||
-        platformNotSupported ||
+        _webContextMenuEnabled ||
         widget.readOnly ||
         _selectionOverlay == null ||
         !_spellCheckResultsReceived ||
@@ -4168,6 +4605,8 @@ class _EditableTextState extends State<_EditableText>
   @override
   String get autofillId => 'EditableText-$hashCode';
 
+  int? _viewId;
+
   @override
   TextInputConfiguration get textInputConfiguration {
     final List<String>? autofillHints =
@@ -4180,7 +4619,9 @@ class _EditableTextState extends State<_EditableText>
           )
         : AutofillConfiguration.disabled;
 
+    _viewId = View.of(context).viewId;
     return TextInputConfiguration(
+      viewId: _viewId,
       inputType: widget.keyboardType,
       readOnly: widget.readOnly,
       obscureText: widget.obscureText,
@@ -4506,6 +4947,15 @@ class _EditableTextState extends State<_EditableText>
   }
 
   void _updateSelection(UpdateSelectionIntent intent) {
+    assert(
+      intent.newSelection.start <= intent.currentTextEditingValue.text.length,
+      'invalid selection: ${intent.newSelection}: it must not exceed the current text length ${intent.currentTextEditingValue.text.length}',
+    );
+    assert(
+      intent.newSelection.end <= intent.currentTextEditingValue.text.length,
+      'invalid selection: ${intent.newSelection}: it must not exceed the current text length ${intent.currentTextEditingValue.text.length}',
+    );
+
     bringIntoView(intent.newSelection.extent);
     userUpdateTextEditingValue(
       intent.currentTextEditingValue.copyWith(selection: intent.newSelection),
@@ -4653,11 +5103,21 @@ class _EditableTextState extends State<_EditableText>
   //   super.build(context); // See AutomaticKeepAliveClientMixin.
 
   //   final TextSelectionControls? controls = widget.selectionControls;
+  //   final TextScaler effectiveTextScaler =
+  //       switch ((widget.textScaler, widget.textScaleFactor)) {
+  //     (final TextScaler textScaler, _) => textScaler,
+  //     (null, final double textScaleFactor) =>
+  //       TextScaler.linear(textScaleFactor),
+  //     (null, null) => MediaQuery.textScalerOf(context),
+  //   };
+
   //   return _CompositionCallback(
   //     compositeCallback: _compositeCallback,
   //     enabled: _hasInputConnection,
   //     child: TextFieldTapRegion(
-  //       onTapOutside: widget.onTapOutside ?? _defaultOnTapOutside,
+  //       groupId: widget.groupId,
+  //       onTapOutside:
+  //           _hasFocus ? widget.onTapOutside ?? _defaultOnTapOutside : null,
   //       debugLabel: kReleaseMode ? null : 'EditableText',
   //       child: MouseRegion(
   //         cursor: widget.mouseCursor ?? SystemMouseCursors.text,
@@ -4666,9 +5126,11 @@ class _EditableTextState extends State<_EditableText>
   //           child: UndoHistory<TextEditingValue>(
   //             value: widget.controller,
   //             onTriggered: (TextEditingValue value) {
-  //               userUpdateTextEditingValue(value, SelectionChangedCause.keyboard);
+  //               userUpdateTextEditingValue(
+  //                   value, SelectionChangedCause.keyboard);
   //             },
-  //             shouldChangeUndoStack: (TextEditingValue? oldValue, TextEditingValue newValue) {
+  //             shouldChangeUndoStack:
+  //                 (TextEditingValue? oldValue, TextEditingValue newValue) {
   //               if (!newValue.selection.isValid) {
   //                 return false;
   //               }
@@ -4694,7 +5156,17 @@ class _EditableTextState extends State<_EditableText>
   //                   break;
   //               }
 
-  //               return oldValue.text != newValue.text || oldValue.composing != newValue.composing;
+  //               return oldValue.text != newValue.text ||
+  //                   oldValue.composing != newValue.composing;
+  //             },
+  //             undoStackModifier: (TextEditingValue value) {
+  //               // On Android we should discard the composing region when pushing
+  //               // a new entry to the undo stack. This prevents the TextInputPlugin
+  //               // from restarting the input on every undo/redo when the composing
+  //               // region is changed by the framework.
+  //               return defaultTargetPlatform == TargetPlatform.android
+  //                   ? value.copyWith(composing: TextRange.empty)
+  //                   : value;
   //             },
   //             focusNode: widget.focusNode,
   //             controller: widget.undoController,
@@ -4702,83 +5174,107 @@ class _EditableTextState extends State<_EditableText>
   //               focusNode: widget.focusNode,
   //               includeSemantics: false,
   //               debugLabel: kReleaseMode ? null : 'EditableText',
-  //               child: Scrollable(
-  //                 key: _scrollableKey,
-  //                 excludeFromSemantics: true,
-  //                 axisDirection: _isMultiline ? AxisDirection.down : AxisDirection.right,
-  //                 controller: _scrollController,
-  //                 physics: widget.scrollPhysics,
-  //                 dragStartBehavior: widget.dragStartBehavior,
-  //                 restorationId: widget.restorationId,
-  //                 // If a ScrollBehavior is not provided, only apply scrollbars when
-  //                 // multiline. The overscroll indicator should not be applied in
-  //                 // either case, glowing or stretching.
-  //                 scrollBehavior: widget.scrollBehavior ?? ScrollConfiguration.of(context).copyWith(
-  //                   scrollbars: _isMultiline,
-  //                   overscroll: false,
-  //                 ),
-  //                 viewportBuilder: (BuildContext context, ViewportOffset offset) {
-  //                   return CompositedTransformTarget(
-  //                     link: _toolbarLayerLink,
-  //                     child: Semantics(
-  //                       onCopy: _semanticsOnCopy(controls),
-  //                       onCut: _semanticsOnCut(controls),
-  //                       onPaste: _semanticsOnPaste(controls),
-  //                       child: _ScribbleFocusable(
-  //                         focusNode: widget.focusNode,
-  //                         editableKey: _editableKey,
-  //                         enabled: widget.scribbleEnabled,
-  //                         updateSelectionRects: () {
-  //                           _openInputConnection();
-  //                           _updateSelectionRects(force: true);
-  //                         },
-  //                         child: _Editable(
-  //                           key: _editableKey,
-  //                           startHandleLayerLink: _startHandleLayerLink,
-  //                           endHandleLayerLink: _endHandleLayerLink,
-  //                           inlineSpan: buildTextSpan(),
-  //                           value: _value,
-  //                           cursorColor: _cursorColor,
-  //                           backgroundCursorColor: widget.backgroundCursorColor,
-  //                           showCursor: _cursorVisibilityNotifier,
-  //                           forceLine: widget.forceLine,
-  //                           readOnly: widget.readOnly,
-  //                           hasFocus: _hasFocus,
-  //                           maxLines: widget.maxLines,
-  //                           minLines: widget.minLines,
-  //                           expands: widget.expands,
-  //                           strutStyle: widget.strutStyle,
-  //                           selectionColor: _selectionOverlay?.spellCheckToolbarIsVisible ?? false
-  //                               ? _spellCheckConfiguration.misspelledSelectionColor ?? widget.selectionColor
-  //                               : widget.selectionColor,
-  //                           textScaler: widget.textScaler ?? MediaQuery.textScalerOf(context),
-  //                           textAlign: widget.textAlign,
-  //                           textDirection: _textDirection,
-  //                           locale: widget.locale,
-  //                           textHeightBehavior: widget.textHeightBehavior ?? DefaultTextHeightBehavior.maybeOf(context),
-  //                           textWidthBasis: widget.textWidthBasis,
-  //                           obscuringCharacter: widget.obscuringCharacter,
-  //                           obscureText: widget.obscureText,
-  //                           offset: offset,
-  //                           rendererIgnoresPointer: widget.rendererIgnoresPointer,
-  //                           cursorWidth: widget.cursorWidth,
-  //                           cursorHeight: widget.cursorHeight,
-  //                           cursorRadius: widget.cursorRadius,
-  //                           cursorOffset: widget.cursorOffset ?? Offset.zero,
-  //                           selectionHeightStyle: widget.selectionHeightStyle,
-  //                           selectionWidthStyle: widget.selectionWidthStyle,
-  //                           paintCursorAboveText: widget.paintCursorAboveText,
-  //                           enableInteractiveSelection: widget._userSelectionEnabled,
-  //                           textSelectionDelegate: this,
-  //                           devicePixelRatio: _devicePixelRatio,
-  //                           promptRectRange: _currentPromptRectRange,
-  //                           promptRectColor: widget.autocorrectionTextRectColor,
-  //                           clipBehavior: widget.clipBehavior,
+  //               child: NotificationListener<ScrollNotification>(
+  //                 onNotification: (ScrollNotification notification) {
+  //                   _handleContextMenuOnScroll(notification);
+  //                   _scribbleCacheKey = null;
+  //                   return false;
+  //                 },
+  //                 child: Scrollable(
+  //                   key: _scrollableKey,
+  //                   excludeFromSemantics: true,
+  //                   axisDirection:
+  //                       _isMultiline ? AxisDirection.down : AxisDirection.right,
+  //                   controller: _scrollController,
+  //                   physics: widget.scrollPhysics,
+  //                   dragStartBehavior: widget.dragStartBehavior,
+  //                   restorationId: widget.restorationId,
+  //                   // If a ScrollBehavior is not provided, only apply scrollbars when
+  //                   // multiline. The overscroll indicator should not be applied in
+  //                   // either case, glowing or stretching.
+  //                   scrollBehavior: widget.scrollBehavior ??
+  //                       ScrollConfiguration.of(context).copyWith(
+  //                         scrollbars: _isMultiline,
+  //                         overscroll: false,
+  //                       ),
+  //                   viewportBuilder:
+  //                       (BuildContext context, ViewportOffset offset) {
+  //                     return CompositedTransformTarget(
+  //                       link: _toolbarLayerLink,
+  //                       child: Semantics(
+  //                         onCopy: _semanticsOnCopy(controls),
+  //                         onCut: _semanticsOnCut(controls),
+  //                         onPaste: _semanticsOnPaste(controls),
+  //                         child: _ScribbleFocusable(
+  //                           focusNode: widget.focusNode,
+  //                           editableKey: _editableKey,
+  //                           enabled: widget.scribbleEnabled,
+  //                           updateSelectionRects: () {
+  //                             _openInputConnection();
+  //                             _updateSelectionRects(force: true);
+  //                           },
+  //                           child: SizeChangedLayoutNotifier(
+  //                             child: _Editable(
+  //                               key: _editableKey,
+  //                               startHandleLayerLink: _startHandleLayerLink,
+  //                               endHandleLayerLink: _endHandleLayerLink,
+  //                               inlineSpan: buildTextSpan(),
+  //                               value: _value,
+  //                               cursorColor: _cursorColor,
+  //                               backgroundCursorColor:
+  //                                   widget.backgroundCursorColor,
+  //                               showCursor: _cursorVisibilityNotifier,
+  //                               forceLine: widget.forceLine,
+  //                               readOnly: widget.readOnly,
+  //                               hasFocus: _hasFocus,
+  //                               maxLines: widget.maxLines,
+  //                               minLines: widget.minLines,
+  //                               expands: widget.expands,
+  //                               strutStyle: widget.strutStyle,
+  //                               selectionColor: _selectionOverlay
+  //                                           ?.spellCheckToolbarIsVisible ??
+  //                                       false
+  //                                   ? _spellCheckConfiguration
+  //                                           .misspelledSelectionColor ??
+  //                                       widget.selectionColor
+  //                                   : widget.selectionColor,
+  //                               textScaler: effectiveTextScaler,
+  //                               textAlign: widget.textAlign,
+  //                               textDirection: _textDirection,
+  //                               locale: widget.locale,
+  //                               textHeightBehavior: widget.textHeightBehavior ??
+  //                                   DefaultTextHeightBehavior.maybeOf(context),
+  //                               textWidthBasis: widget.textWidthBasis,
+  //                               obscuringCharacter: widget.obscuringCharacter,
+  //                               obscureText: widget.obscureText,
+  //                               offset: offset,
+  //                               rendererIgnoresPointer:
+  //                                   widget.rendererIgnoresPointer,
+  //                               cursorWidth: widget.cursorWidth,
+  //                               cursorHeight: widget.cursorHeight,
+  //                               cursorRadius: widget.cursorRadius,
+  //                               cursorOffset:
+  //                                   widget.cursorOffset ?? Offset.zero,
+  //                               selectionHeightStyle:
+  //                                   widget.selectionHeightStyle,
+  //                               selectionWidthStyle: widget.selectionWidthStyle,
+  //                               paintCursorAboveText:
+  //                                   widget.paintCursorAboveText,
+  //                               enableInteractiveSelection:
+  //                                   widget._userSelectionEnabled,
+  //                               textSelectionDelegate: this,
+  //                               devicePixelRatio: _devicePixelRatio,
+  //                               promptRectRange: _currentPromptRectRange,
+  //                               promptRectColor:
+  //                                   widget.autocorrectionTextRectColor,
+  //                               clipBehavior: widget.clipBehavior,
+  //                             ),
+  //                           ),
   //                         ),
   //                       ),
-  //                     ),
-  //                   );
-  //                 },
+  //                     );
+  //                   },
+  //                 ),
   //               ),
   //             ),
   //           ),
@@ -5210,11 +5706,7 @@ class _ScribblePlaceholder extends WidgetSpan {
     if (hasStyle) {
       builder.pushStyle(style!.getTextStyle(textScaler: textScaler));
     }
-    builder.addPlaceholder(
-      size.width * textScaler.textScaleFactor,
-      size.height * textScaler.textScaleFactor,
-      alignment,
-    );
+    builder.addPlaceholder(size.width, size.height, alignment);
     if (hasStyle) {
       builder.pop();
     }
@@ -5520,7 +6012,7 @@ class _UpdateTextSelectionVerticallyAction<
     final TextPosition newExtent = shouldMove
         ? currentRun.current
         : intent.forward
-            ? TextPosition(offset: state._value.text.length)
+            ? TextPosition(offset: value.text.length)
             : const TextPosition(offset: 0);
     final TextSelection newSelection = collapseSelection
         ? TextSelection.fromPosition(newExtent)
@@ -5581,21 +6073,6 @@ class _CopySelectionAction extends ContextAction<CopySelectionTextIntent> {
   @override
   bool get isActionEnabled =>
       state._value.selection.isValid && !state._value.selection.isCollapsed;
-}
-
-/// The start and end glyph heights of some range of text.
-@immutable
-class _GlyphHeights {
-  const _GlyphHeights({
-    required this.start,
-    required this.end,
-  });
-
-  /// The glyph height of the first line.
-  final double start;
-
-  /// The glyph height of the last line.
-  final double end;
 }
 
 /// A [ClipboardStatusNotifier] whose [value] is hardcoded to
