@@ -77,19 +77,18 @@ class _SpellCheckConfiguration {
 
   @override
   String toString() {
-    return '''
-  spell check enabled   : $_spellCheckEnabled
-  spell check service   : $spellCheckService
-  misspelled text style : $misspelledTextStyle
-  spell check suggestions toolbar builder: $spellCheckSuggestionsToolbarBuilder
-'''
-        .trim();
+    return '${objectRuntimeType(this, 'SpellCheckConfiguration')}('
+        '${_spellCheckEnabled ? 'enabled' : 'disabled'}, '
+        'service: $spellCheckService, '
+        'text style: $misspelledTextStyle, '
+        'toolbar builder: $spellCheckSuggestionsToolbarBuilder'
+        ')';
   }
 
   @override
   bool operator ==(Object other) {
-    if (identical(this, other)) {
-      return true;
+    if (other.runtimeType != runtimeType) {
+      return false;
     }
 
     return other is _SpellCheckConfiguration &&
@@ -130,7 +129,8 @@ List<SuggestionSpan> _correctSpellCheckResults(
     final int spanLength = currentSpan.range.end - currentSpan.range.start;
 
     // Try finding SuggestionSpan from resultsText in new text.
-    final RegExp currentSpanTextRegexp = RegExp('\\b$currentSpanText\\b');
+    final String escapedText = RegExp.escape(currentSpanText);
+    final RegExp currentSpanTextRegexp = RegExp('\\b$escapedText\\b');
     final int foundIndex =
         newText.substring(searchStart).indexOf(currentSpanTextRegexp);
 
@@ -142,7 +142,7 @@ List<SuggestionSpan> _correctSpellCheckResults(
     final bool currentSpanFoundElsewhere = foundIndex >= 0;
 
     if (currentSpanFoundExactly || currentSpanFoundExactlyWithOffset) {
-      // currentSpan was found at the same index in newText and resutsText
+      // currentSpan was found at the same index in newText and resultsText
       // or at the same index with the previously calculated adjustment by
       // the offset value, so apply it to new text by adding it to the list of
       // corrected results.
@@ -206,7 +206,7 @@ TextSpan buildTextSpanWithSpellCheckSuggestions(
 
   // We will draw the TextSpan tree based on the composing region, if it is
   // available.
-  // TODO(camsim99): The two separate stratgies for building TextSpan trees
+  // TODO(camsim99): The two separate strategies for building TextSpan trees
   // based on the availability of a composing region should be merged:
   // https://github.com/flutter/flutter/issues/124142.
   final bool shouldConsiderComposingRegion =
